requirements.txt
 #projenin calısması için gerekli kutuphaneler 
Flask==3.0.0
psycopg2-binary
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Flask-CORS==4.0.0